upgrade auto catalog

+++ src/Services/AutoCatalog/AutoCatalog.Application/Cars/CreateCar/CreateCarEndpoint.cs
@@ -21,8 +21,8 @@ public record CreateCarRequest(
     int YearFrom,
     int YearTo,
     Guid PhotoId,
-    float Consumption,
-    float Acceleration,
+    decimal Consumption,
+    decimal Acceleration,
     int FuelTankCapacity,
     DimensionsDto DimensionsDto);
